+++ src/main/java/cv/unipiaget/manutencao_aeronave/Controller/PecaController.java
@@ -8,7 +8,9 @@ package cv.unipiaget.manutencao_aeronave.Controller;
 
 import cv.unipiaget.manutencao_aeronave.Entities.PecaEntity;
 import cv.unipiaget.manutencao_aeronave.Services.PecaService;
+import cv.unipiaget.manutencao_aeronave.recursos.ErrorResponse;
 import org.springframework.beans.factory.annotation.Autowired;
+import org.springframework.http.HttpEntity;
 import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -34,15 +36,22 @@ public class PecaController {
     }
 
     @GetMapping("/{id}")
-    public ResponseEntity<PecaEntity> obterPeca(@PathVariable int id) {
-        Optional<PecaEntity> peca = pecaService.obterPecaPorId(id);
-        return peca.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
+    public HttpEntity<? extends Object> obterPeca(@PathVariable int id) {
+
+        Optional<PecaEntity> pecaExistente = pecaService.obterPecaPorId(id);
+
+        if (pecaExistente.isPresent()) {
+            return pecaExistente.map(ResponseEntity::ok).orElseGet(() -> ResponseEntity.notFound().build());
+        } else {
+            String mensagem = "Peca com ID " + id + " não encontrada.";
+            ErrorResponse errorResponse = new ErrorResponse(HttpStatus.NOT_FOUND, mensagem);
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorResponse);
+        }
     }
 
     @PostMapping
     public ResponseEntity<PecaEntity> adicionarPeca(@RequestBody PecaEntity peca) {
         try {
-            // guardar na base de dados
             PecaEntity pecaSalva = pecaService.salvarPeca(peca);
 
             return ResponseEntity.status(HttpStatus.CREATED).body(pecaSalva);
@@ -53,20 +62,36 @@ public class PecaController {
 
 
     @PutMapping("/{id}")
-    public ResponseEntity<PecaEntity> atualizarPeca(@PathVariable int id, @RequestBody PecaEntity peca) {
+    public ResponseEntity<?> atualizarPeca(@PathVariable int id, @RequestBody PecaEntity peca) {
         Optional<PecaEntity> pecaExistente = pecaService.obterPecaPorId(id);
+
         if (pecaExistente.isPresent()) {
             peca.setId(id);
             PecaEntity pecaAtualizada = pecaService.salvarPeca(peca);
-            return ResponseEntity.ok(pecaAtualizada);
+            String mensagem = "Peca com ID " + id + " atualizada com sucesso.";
+            return ResponseEntity.ok().body(mensagem);
         } else {
-            return ResponseEntity.notFound().build();
+            String mensagem = "Peca com ID " + id + " não encontrada. A atualização não pode ser realizada.";
+            ErrorResponse errorResponse = new ErrorResponse(HttpStatus.NOT_FOUND, mensagem);
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorResponse);
         }
     }
+
     @DeleteMapping("/{id}")
-    public ResponseEntity<Void> excluirPeca(@PathVariable int id) {
-        pecaService.excluirPeca(id);
-        return ResponseEntity.noContent().build();
+    public ResponseEntity<?> excluirPeca(@PathVariable int id) {
+        Optional<PecaEntity> pecaExistente = pecaService.obterPecaPorId(id);
+
+        if (pecaExistente.isPresent()) {
+            pecaService.excluirPeca(id);
+            String mensagem = "Peca com ID " + id + " excluída com sucesso.";
+            return ResponseEntity.ok().body(mensagem);
+        } else {
+            String mensagem = "Peca com ID " + id + " não encontrada.";
+            ErrorResponse errorResponse = new ErrorResponse(HttpStatus.NOT_FOUND, mensagem);
+            return ResponseEntity.status(HttpStatus.NOT_FOUND).body(errorResponse);
+        }
     }
+
+
 
 }
